Code
This is the VEX robot X Drive code. The motor config is shown in the code, I accidentally reversed the motor's spins, but it works so I decided not to change it. The Diagonals still need to be coded, along with the autonomous, and the other robot functions such as the intake, but for now this is just the basic x-drive.